requirements.txt
fastapi
uvicorn[standard]
pydantic
GitPython
requests
python-multipart
jinja2
apscheduler
aiofiles
xmltodict
